Changes to event manager and event details
Added frontend pages for event admin and event details.
Changed url path to match new files.
Adjusted Session variables to match those from login.
Changed Id in the DB insert command
